% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test_625.R
\name{cor_test_625}
\alias{cor_test_625}
\title{Title}
\usage{
cor_test_625(x, y, method = "pearson")
}
\arguments{
\item{x}{A numerical vector}

\item{y}{A numerical vector}

\item{method}{Method used to calculate correlation value}
}
\value{
A numerical vector of one containing the correlation value
}
\description{
Title
}
\examples{
x <- c(6, 12, 13, 17, 22, 25, 27, 29, 30, 32)
y <- c(45, 47, 39, 58, 68, 76, 75, 74, 78, 81)
cor_test_625(x, y, method = "pearson")
}
