% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test_625.R
\name{cortest625}
\alias{cortest625}
\title{Correlation Package}
\usage{
cortest625(x, y, method = "pearson")
}
\arguments{
\item{x}{A numerical vector.}

\item{y}{A numerical vector.}

\item{method}{Method used to calculate correlation value. Methods include Pearson, Spearman, Kendall, All. Default is Pearson.}
}
\value{
A numerical vector of one containing the correlation value (All creates a 1x3 data frame containing all 3 values).
}
\description{
Correlation Package
}
\examples{
Creating two numerical vectors
x <- c(6, 12, 13, 17, 22, 25, 27, 29, 30, 32)
y <- c(45, 47, 39, 58, 68, 76, 75, 74, 78, 81)
cortest625(x, y, method = "pearson")

Using the mtcars dataset
cortest625(mtcars$mpg, mtcars$drat, method = "spearman")

Randomly generating data
cortest625(rnorm(100), rnorm(100), method = "kendall")
}
